create interface service  Commende et class service commende

+++ DreamCakes/src/main/java/com/cakes/Repository/CommendeRepository.java
@@ -3,5 +3,5 @@ package com.cakes.Repository;
 import com.cakes.Model.Commende;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface CommendeRepository extends JpaRepository<Commende, Integer> {
+public interface CommendeRepository extends JpaRepository<Commende, Long> {
 }
